[examples] polymer_crankshaft.rs splitted.
