+++ Cargo.toml
@@ -43,6 +43,7 @@ name = "surpass_alfa"
 path = "src/surpass_alfa.rs"
 
 
+
 [package.metadata.docs.rs]
 # To build locally use
 #   RUSTDOCFLAGS="--html-in-header katex-header.html" cargo doc --no-deps --document-private-items --open
