+++ examples/argon.rs
@@ -3,9 +3,9 @@ use std::time::Instant;
 use clap::Parser;
 
 use bioshell_cartesians::movers::{ChangeVolume, SingleAtomMove};
-use bioshell_cartesians::r_end_squared::PdbTrajectory;
+use bioshell_cartesians::pdb_trajectory::PdbTrajectory;
 use bioshell_cartesians::{
-    compute_box_width, coordinates_to_pdb, cubic_grid_atoms, ArgonRules, CartesianSystem, Coordinates,
+    compute_box_width, write_coordinates_to_pdb, cubic_grid_atoms, ArgonRules, CartesianSystem, Coordinates,
     NbList,
 };
 use bioshell_ff::nonbonded::{LennardJonesHomogenic, PairwiseNonbondedEvaluator};
@@ -72,7 +72,7 @@ pub fn main() {
     let mut coords = Coordinates::new(n_atoms);
     coords.set_box_len(compute_box_width(SIGMA as f64, n_atoms, density));
     cubic_grid_atoms(&mut coords);
-    coordinates_to_pdb(&coords, 1, tra_fname.as_str(), false);
+    write_coordinates_to_pdb(&coords, 1, tra_fname.as_str(), false);
 
     // ---------- Create system's list of neighbors
     let nbl: NbList = NbList::new(CUTOFF as f64, 4.0, Box::new(ArgonRules {}));
@@ -114,5 +114,5 @@ pub fn main() {
         &mut observers,
     );
 
-    coordinates_to_pdb(&system.get_coordinates(), 1, final_fname.as_str(), false);
+    write_coordinates_to_pdb(&system.get_coordinates(), 1, final_fname.as_str(), false);
 }
